# harness/specs/001-fastapi-openvino-inference/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OpenVINO Model Server API
  version: 0.1.0
  description: Geração de texto por FastAPI usando OVMS na rede interna.
servers:
  - url: /
paths:
  /healthz:
    get:
      operationId: healthz
      responses:
        '200':
          description: API viva
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
  /readyz:
    get:
      operationId: readyz
      responses:
        '200':
          description: Modelo pronto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadinessResponse' }
        '503':
          description: Modelo indisponível ou configuração inválida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadinessResponse' }
  /v1/generate/sync:
    post:
      operationId: generateSync
      requestBody: { $ref: '#/components/requestBodies/GenerationRequest' }
      responses: { '200': { $ref: '#/components/responses/GenerationResponse' }, '4XX': { $ref: '#/components/responses/ErrorResponse' }, '5XX': { $ref: '#/components/responses/ErrorResponse' } }
  /v1/generate/async:
    post:
      operationId: generateAsync
      requestBody: { $ref: '#/components/requestBodies/GenerationRequest' }
      responses: { '200': { $ref: '#/components/responses/GenerationResponse' }, '4XX': { $ref: '#/components/responses/ErrorResponse' }, '5XX': { $ref: '#/components/responses/ErrorResponse' } }
  /v1/generate/stream:
    post:
      operationId: generateStream
      requestBody: { $ref: '#/components/requestBodies/GenerationRequest' }
      responses:
        '200':
          description: Eventos SSE incrementais
          content:
            text/event-stream:
              schema:
                type: string
                description: Cada evento contém `event` e JSON em `data`.
        '4XX': { $ref: '#/components/responses/ErrorResponse' }
        '5XX': { $ref: '#/components/responses/ErrorResponse' }
components:
  requestBodies:
    GenerationRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenerationRequest' }
  responses:
    GenerationResponse:
      description: Geração concluída
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenerationResponse' }
    ErrorResponse:
      description: Erro controlado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    GenerationRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 12000 }
        max_tokens: { type: integer, minimum: 1, maximum: 512, default: 128 }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.2 }
    GenerationResponse:
      type: object
      required: [request_id, model, text, finish_reason]
      properties:
        request_id: { type: string }
        model: { type: string }
        text: { type: string }
        finish_reason: { type: string }
        usage: { type: object, additionalProperties: true, nullable: true }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, enum: [invalid_request, capacity_limited, not_ready, upstream_unavailable, upstream_failed, generation_timeout] }
            message: { type: string }
            request_id: { type: string }
    HealthResponse:
      type: object
      required: [status]
      properties: { status: { type: string, enum: [alive] } }
    ReadinessResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ready, not_ready] }
        reason: { type: string, nullable: true }